requirements.txt
# PipLine Treasury System - Development Requirements
# Python Version Requirement: >=3.9,<3.14
# Core Flask Dependencies
flask==3.1.1
flask-cors==6.0.1  # Latest stable (no 6.1.0)
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
flask-migrate==4.0.7
flask-limiter==3.8.0
flask-jwt-extended==4.7.1
wtforms==3.2.1
werkzeug==3.1.3

# Flask Core Dependencies (explicitly listed for compatibility)
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
blinker==1.9.0

# Compression Dependencies (for flask-compress)
brotli==1.2.0
backports.zstd==1.5.6.1; python_version < "3.14"

# JWT Dependencies (for flask-jwt-extended)
PyJWT==2.10.1
cryptography==44.0.0

# Database Dependencies
sqlalchemy==2.0.36
psycopg2-binary==2.9.11  # PostgreSQL adapter
psycopg2-pool==1.1  # Connection pooling for PostgreSQL
pyodbc==5.1.0  # MSSQL Server driver (ODBC)

# Data Processing
# Note: pandas 2.2.3 may have compatibility issues with Python 3.13
# Consider using pandas>=2.2.0 or later versions if available
pandas==2.2.3
openpyxl==3.1.5
numpy>=1.24.0; python_version < "3.13"

# Exchange Rate Dependencies
yfinance==0.2.50  # Yahoo Finance API for exchange rates (stable version)
# yfinance dependencies
lxml>=4.9.0
html5lib>=1.1
beautifulsoup4>=4.12.0
multitasking>=0.0.11
pytz>=2023.3
pandas-datareader>=0.10.0

# Visualization and Analytics
plotly==5.24.1
matplotlib==3.9.3
seaborn==0.13.2
# Plotly dependencies
kaleido>=0.2.1
# Matplotlib dependencies
Pillow>=10.0.0
cycler>=0.12.0
fonttools>=4.47.0
kiwisolver>=1.4.5
packaging>=23.2
pyparsing>=3.1.1

# Caching and Session Management
redis==5.2.1
flask_session==0.8.0
flask-compress==1.23
celery==5.4.0
flower==2.0.1
# Celery dependencies
kombu==5.4.2
vine==5.1.0
billiard==4.2.1

# Security and Validation
email-validator==2.2.0
filetype==1.2.0
flask-talisman==1.1.0
bleach==6.2.0

# Real-time Features
flask-socketio==5.4.1
# eventlet==0.40.3  # Disabled due to Python 3.13 compatibility issues
# Using threading mode instead of eventlet for Python 3.13 compatibility
python-socketio==5.11.4
python-engineio==4.10.1

# Enhanced Services
requests==2.32.3  # Security fixes
# kafka-python is unmaintained - replaced with confluent-kafka-python
confluent-kafka==2.6.1

# AI / ChatGPT Integration
openai==1.57.4  # CRITICAL UPDATE - major version changes

# Internationalization
flask-babel==4.0.0
babel==2.16.0

# System Monitoring
psutil==6.1.1

# Logging (Windows-compatible)
concurrent-log-handler==0.9.25  # Windows-safe log rotation

# Background Tasks
schedule==1.2.2

# Environment Management
python-dotenv==1.0.1

# Database Recovery and Maintenance
# Note: sqlite3 is included in Python standard library
# All database recovery functionality uses standard library modules

# Production Server (Required for Production)
gunicorn==23.0.0  # WSGI server for production deployment

# Security Auditing Tools (Recommended)
pip-audit==2.7.3  # Security vulnerability scanner
# safety==3.2.11  # Alternative security checker

# Development Tools (Optional - Uncomment for development)
# black==24.10.0  # Code formatting
# flake8==7.1.1  # Linting
# pylint==3.3.3  # Advanced linting
# mypy==1.13.0  # Type checking
# isort==5.13.2  # Import sorting

# Testing Tools
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
pytest-mock==3.14.0
faker==33.1.0

# Monitoring & Error Tracking
sentry-sdk[flask]==2.20.0  # Error tracking and monitoring
prometheus-flask-exporter==0.23.1  # Metrics export for monitoring
prometheus-client==0.20.0  # Prometheus Python client

# Backup & Cloud Storage (Optional - for S3 backups)
boto3==1.35.82  # AWS SDK for S3 backups